refactoring panels and react runner into seperate components
pass revision to parent

new comment
