+++ app/javascripts/constants/frame_content.js
@@ -8,8 +8,7 @@ export default (cssResources, jsResources) =>`
 <div id="client_results"></div>
 <script src="/javascripts/react.js"></script>
 <script src="/javascripts/reactdom.js"></script>
-<div id="injected-scripts">
-    ${jsResources.map(r =>'<script type="text/javascript" src="' + r + '"></script>')}
+<div id="injected-scripts">${jsResources.map(r =>'<script type="text/javascript" src="' + r + '"></script>')}
 </div>
 </body>
 </html>
